Restructure default project
